backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
greenlet==3.1.1
asyncpg==0.30.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
pwdlib[argon2]==0.2.1
httpx==0.28.1
beautifulsoup4==4.12.3
email-validator==2.2.0
python-multipart==0.0.20
